solution/requirements.txt
psycopg2-binary
django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
requests==2.31